+++ cpp/src/sdk/MiraiCPNewThread.cpp
@@ -16,7 +16,7 @@
 
 #include "MiraiCPNewThread.h"
 #include "Event.h"
-
+#include "Events/MiraiCPExceptionEvent.h"
 #include <ostream>